[feat] add TemplateEngine interface
